@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016. Vijai Chandra Prasad R.
+ * Copyright (c) 2016-2018. Vijai Chandra Prasad R.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -18,15 +18,15 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion "27.0.3"
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
     defaultConfig {
         applicationId "com.orpheusdroid.screenrecorder"
-        minSdkVersion 21
-        targetSdkVersion 27
-        versionCode 24
-        versionName "1.8.7"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        minSdkVersion 22
+        targetSdkVersion 28
+        versionCode 33
+        versionName "2.0.3"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     lintOptions {
         abortOnError false
@@ -57,6 +57,17 @@ android {
             dimension "tier"
             versionNameSuffix '-fdroid'
         }
+        playstoreDonate {
+            dimension "tier"
+            versionNameSuffix '-playstore Donate'
+            applicationIdSuffix '.donate'
+            buildConfigField "String", "APP_PUB_KEY", getAppKey("app_pub_key_donate")
+        }
+    }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
 }
 
@@ -79,18 +90,22 @@ if (project.hasProperty('RELEASE_STORE_FILE')) {
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.android.support.constraint:constraint-layout:1.0.2'
-    androidTestImplementation('com.android.support.test.espresso:espresso-core:2.2.2', {
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.0-beta1'
+    androidTestImplementation('androidx.test.espresso:espresso-core:3.1.0-alpha4', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support:design:27.1.1'
-    implementation 'com.android.support:cardview-v7:27.1.1'
-    implementation 'com.android.support:support-v13:27.1.1'
-    implementation 'ly.count.android:sdk:17.09.2'
+    implementation 'androidx.appcompat:appcompat:1.0.2'
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v13:1.0.0'
+    implementation 'ly.count.android:sdk:18.08.1'
     implementation 'life.knowledge4:k4l-video-trimmer:1.0'
+    implementation 'com.github.google:cameraview:3eaeac09a7f'
+    implementation 'com.github.MFlisar:changelog:1.1.2'
+    implementation 'com.github.topjohnwu.libsu:core:2.5.0'
 
     playstoreImplementation 'org.solovyev.android:checkout:1.2.1'
+    playstoreDonateImplementation 'org.solovyev.android:checkout:1.2.1'
 
     testImplementation 'junit:junit:4.12'
 }
