@@ -21,9 +21,9 @@ import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.CardView;
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.cardview.widget.CardView;
 import android.view.MenuItem;
 import android.view.View;
 
@@ -34,6 +34,9 @@ public class DonateActivity extends AppCompatActivity implements View.OnClickLis
         String theme = PreferenceManager.getDefaultSharedPreferences(this)
                 .getString(getString(R.string.preference_theme_key), Const.PREFS_LIGHT_THEME);
         switch (theme) {
+            case Const.PREFS_WHITE_THEME:
+                setTheme(R.style.AppTheme_White);
+                break;
             case Const.PREFS_DARK_THEME:
                 setTheme(R.style.AppTheme_Dark);
                 break;
