@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2017. Vijai Chandra Prasad R.
+ * Copyright (c) 2016-2018. Vijai Chandra Prasad R.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -23,13 +23,6 @@ import android.content.Intent;
 import android.graphics.drawable.ColorDrawable;
 import android.net.Uri;
 import android.preference.PreferenceManager;
-import android.support.v4.content.ContextCompat;
-import android.support.v4.content.FileProvider;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.view.ActionMode;
-import android.support.v7.widget.PopupMenu;
-import android.support.v7.widget.RecyclerView;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.Menu;
@@ -45,10 +38,10 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import com.orpheusdroid.screenrecorder.Const;
-import com.orpheusdroid.screenrecorder.EditVideoActivity;
 import com.orpheusdroid.screenrecorder.R;
-import com.orpheusdroid.screenrecorder.VideosListFragment;
 import com.orpheusdroid.screenrecorder.encoder.Mp4toGIFConverter;
+import com.orpheusdroid.screenrecorder.ui.EditVideoActivity;
+import com.orpheusdroid.screenrecorder.ui.VideosListFragment;
 
 import java.io.File;
 import java.text.SimpleDateFormat;
@@ -56,6 +49,14 @@ import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.Locale;
+
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.view.ActionMode;
+import androidx.appcompat.widget.PopupMenu;
+import androidx.core.content.ContextCompat;
+import androidx.core.content.FileProvider;
+import androidx.recyclerview.widget.RecyclerView;
 
 /**
  * <p>
@@ -435,9 +436,9 @@ public class VideoRecyclerAdapter extends RecyclerView.Adapter<RecyclerView.View
      */
     private void deleteVideos(ArrayList<Video> deleteVideos) {
         for (Video video : deleteVideos) {
-            if (video.getFile().delete()) {
+            if (!video.isSection() && video.getFile().delete()) {
                 notifyItemRemoved(videos.indexOf(video));
-                videos.remove(videos.indexOf(video));
+                videos.remove(video);
             }
         }
         notifyDataSetChanged();
