@@ -18,13 +18,8 @@
 package com.orpheusdroid.screenrecorder;
 
 import android.content.Intent;
-import android.net.Uri;
-import android.preference.PreferenceManager;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.CardView;
-import android.util.Log;
+import android.preference.PreferenceManager;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.Toast;
@@ -39,7 +34,11 @@ import org.solovyev.android.checkout.Purchase;
 
 import javax.annotation.Nonnull;
 
-public class DonateActivity extends AppCompatActivity implements View.OnClickListener{
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.cardview.widget.CardView;
+
+public class DonateActivity extends AppCompatActivity implements View.OnClickListener {
 
     private final ActivityCheckout mCheckout = Checkout.forActivity(this, ScreenCamApp.get().getBilling());
     private Inventory mInventory;
@@ -49,7 +48,10 @@ public class DonateActivity extends AppCompatActivity implements View.OnClickLis
 
         String theme = PreferenceManager.getDefaultSharedPreferences(this)
                 .getString(getString(R.string.preference_theme_key), Const.PREFS_LIGHT_THEME);
-        switch (theme){
+        switch (theme) {
+            case Const.PREFS_WHITE_THEME:
+                setTheme(R.style.AppTheme_White);
+                break;
             case Const.PREFS_DARK_THEME:
                 setTheme(R.style.AppTheme_Dark);
                 break;
@@ -68,17 +70,15 @@ public class DonateActivity extends AppCompatActivity implements View.OnClickLis
         CardView donate1 = findViewById(R.id.donate_1);
         CardView donate2 = findViewById(R.id.donate_2);
         CardView donate5 = findViewById(R.id.donate_5);
-        CardView donatePayPal = findViewById(R.id.donate_paypal);
 
         donate1.setOnClickListener(this);
         donate2.setOnClickListener(this);
         donate5.setOnClickListener(this);
-        donatePayPal.setOnClickListener(this);
 
         setupPurchase();
     }
 
-    private void setupPurchase(){
+    private void setupPurchase() {
         mCheckout.start();
 
         mCheckout.createPurchaseFlow(new PurchaseListener());
@@ -89,7 +89,7 @@ public class DonateActivity extends AppCompatActivity implements View.OnClickLis
                 .loadSkus(ProductTypes.IN_APP, "sku_1", "sku_2", "sku_5"), new InventoryCallback());
     }
 
-    private void makePurchase(final String sku){
+    private void makePurchase(final String sku) {
         mCheckout.whenReady(new Checkout.EmptyListener() {
             @Override
             public void onReady(BillingRequests requests) {
@@ -98,7 +98,7 @@ public class DonateActivity extends AppCompatActivity implements View.OnClickLis
         });
     }
 
-    private void consumePurchase(final Purchase purchase){
+    private void consumePurchase(final Purchase purchase) {
         mCheckout.whenReady(new Checkout.EmptyListener() {
             @Override
             public void onReady(@Nonnull BillingRequests requests) {
@@ -133,7 +133,7 @@ public class DonateActivity extends AppCompatActivity implements View.OnClickLis
 
     @Override
     public void onClick(View view) {
-        switch (view.getId()){
+        switch (view.getId()) {
             case R.id.donate_1:
                 makePurchase("sku_1");
                 break;
@@ -142,12 +142,6 @@ public class DonateActivity extends AppCompatActivity implements View.OnClickLis
                 break;
             case R.id.donate_5:
                 makePurchase("sku_5");
-                break;
-            case R.id.donate_paypal:
-                String url = "http://paypal.me/vijaichander/5";
-                Intent donateURL = new Intent(Intent.ACTION_VIEW);
-                donateURL.setData(Uri.parse(url));
-                startActivity(donateURL);
                 break;
         }
     }
@@ -162,7 +156,6 @@ public class DonateActivity extends AppCompatActivity implements View.OnClickLis
     private class PurchaseListener extends EmptyRequestListener<Purchase> {
         @Override
         public void onSuccess(Purchase purchase) {
-            // here you can process the loaded purchase
             consumePurchase(purchase);
         }
 
@@ -176,7 +169,12 @@ public class DonateActivity extends AppCompatActivity implements View.OnClickLis
     private class InventoryCallback implements Inventory.Callback {
         @Override
         public void onLoaded(Inventory.Products products) {
-            // your code here
+            // Make sure to consume all previous purchases
+            for (Inventory.Product product : products) {
+                for (Purchase purchase : product.getPurchases()) {
+                    consumePurchase(purchase);
+                }
+            }
         }
     }
 }
